requirements.txt
openai
langchain
streamlit
python-dotenv
youtube-transcript-api
tiktoken
faiss-cpu